REFACTOR: Store info about document ref in `LiveDoc`.
